Adds Stripe checkout button and form
Copies simple form from the Stripe Checkout documentation. Adds comments
explaining all of the relevant options for the form. Will be modifying
to communicate with the server correctly and customize for our needs.
To use test, use 4242 4242 4242 4242 as the card number, any future date
as the expiration, and any random numbers for CSV and zip code.