Sets up protocol for guest user
On document ready, a guest user is created and stored in app.user.
When a user signs up, the user's information overwrites the guest's
information and the account is no longer a guest account. On sign in,
the user's past information overwrites the guest account. On sign out,
a new guest user account is created and the guest account is now stored
in app.user